Changed ID type from Integer to Long.

+++ src/main/java/pl/beng/thesis/repository/AddressRepository.java
@@ -3,5 +3,5 @@ package pl.beng.thesis.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.beng.thesis.model.Address;
 
-public interface AddressRepository extends JpaRepository<Address, Integer> {
+public interface AddressRepository extends JpaRepository<Address, Long> {
 }
