+++ src/main/java/pl/beng/thesis/repository/EducationRepository.java
@@ -3,5 +3,5 @@ package pl.beng.thesis.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.beng.thesis.model.Education;
 
-public interface EducationRepository extends JpaRepository<Education, Integer> {
+public interface EducationRepository extends JpaRepository<Education, Long> {
 }
